pyproject.toml
[project]
name = "promptlab"
version = "0.1.0"
description = "Prompt testing framework — pytest for LLM prompts"
readme = "README.md"
authors = [{name = "Donald Havery", email = "dbhavery@gmail.com"}]
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["llm", "prompt", "testing", "cli", "ai", "evaluation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "click==8.1.8",
    "pyyaml==6.0.2",
    "httpx==0.28.1",
    "rich==13.9.4",
]

[project.optional-dependencies]
claude = ["anthropic==0.52.0"]
gemini = ["google-genai==1.14.0"]
all = ["anthropic==0.52.0", "google-genai==1.14.0"]
dev = ["pytest==8.3.5", "pytest-asyncio==0.26.0"]

[project.scripts]
promptlab = "promptlab.cli:main"

[project.urls]
Homepage = "https://github.com/dbhavery/promptlab"
Repository = "https://github.com/dbhavery/promptlab"
Issues = "https://github.com/dbhavery/promptlab/issues"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["promptlab*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
